refactor A*
Previously it was crapped together, now it follows the exact same implementation details of BFS so you can see the very minimal differences between the algorithms.  Another notable change is that the A* algorithm now has the heuristic injected so it can be easily swapped out for different runs.

The path cost is now added to the queue as part of the state.  This makes it a little more clean than trying to use the open list key as the path cost which isn’t true for A* and derivatives.